refactor: simplify stupid handler logic
