--- src/routes/api/post.json.ts
@@ -1,5 +1,0 @@
-import { getResourcesAsync } from "$lib/utils/fetch-data";
-import { getHandler } from "$lib/utils/handler";
-import type { RequestHandler } from "@sveltejs/kit";
-
-export const get = getHandler(() => getResourcesAsync("post")) as RequestHandler;
